Rename MagicMethod to Event Method
