+++ ruby/examples/slider.rb
@@ -2,7 +2,7 @@ require 'skatolo'
 # In this simple sketch we attach a slider to skatolo in the regular way, with
 # a named slider 'background_color' and thanks to some fancy metaprogramming
 # we can read the result from background_color_value
-include MagicMethod
+include EventMethod
 attr_reader :skatolo
 
 def settings
