+++ ruby/lib/skatolo.rb
@@ -1,4 +1,4 @@
 require 'jruby/core_ext'
 require_relative 'skatolo.jar'
-require_relative '../lib/skatolo/skatolo'
-require_relative '../lib/skatolo/magic_method'
+require_relative 'skatolo/skatolo'
+require_relative 'skatolo/event_method'
