+++ ruby/lib/skatolo/event_method.rb
